extract import / export panel to its own component
